% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{create_miRTalk}
\alias{create_miRTalk}
\title{Create miRTalk object}
\usage{
create_miRTalk(sc_data, sc_celltype, species)
}
\arguments{
\item{sc_data}{A data.frame or matrix or dgCMatrix containing raw counts of single-cell RNA-seq data. see \code{\link{demo_sc_data}}}

\item{sc_celltype}{A character containing the cell type of the single-cell RNA-seq data}

\item{species}{A character meaning species of the single-cell transcriptomics data.\code{'Human'}, \code{'Mouse'} or \code{'Rat'}}
}
\value{
miRTalk object
}
\description{
create miRTalk object using single-cell transcriptomics data
}
