% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_gene2go}
\alias{demo_gene2go}
\title{Demo data of gene2go}
\usage{
demo_gene2go()
}
\description{
Demo data of gene2go
}
\details{
\code{gene2go} must be a \code{data.frame} object with three columns, namely \code{'symbol'}, \code{'GO_term'}, \code{'species'}.
}
\examples{
gene2go_demo <- demo_gene2go()
}
