% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_mir2tar}
\alias{demo_mir2tar}
\title{Demo data of mir2tar}
\usage{
demo_mir2tar()
}
\description{
Demo data of mir2tar
}
\details{
\code{mir2tar} must be a \code{data.frame} object with four columns, namely \code{'miRNA'}, \code{'miRNA_mature'}, \code{'target_gene'}, \code{'species'}
}
\examples{
mir2tar_demo <- demo_mir2tar()
}
