% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_mir_info}
\alias{demo_mir_info}
\title{Demo data of mir_info}
\usage{
demo_mir_info()
}
\description{
Demo data of mir_info
}
\details{
\code{mir_info} must be a \code{data.frame} object with four columns, namely \code{'miRNA'}, \code{'miRNA_mature'}, \code{'gene'}, \code{'species'}
}
\examples{
mir_info_demo <- demo_mir_info()
}
