% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_pathways}
\alias{demo_pathways}
\title{Demo data of pathways}
\usage{
demo_pathways()
}
\description{
Demo data of pathways
}
\details{
\code{pathways} must be a \code{data.frame} object with four columns, namely \code{'src'}, \code{'dest'}, \code{'pathway'}, \code{'species'}
}
\examples{
pathways_demo <- demo_pathways()
}
