% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{evbiog}
\alias{evbiog}
\title{evbiog}
\format{
An object of class \code{data.frame} with 75 rows and 2 columns.
}
\source{
\url{https://www.gsea-msigdb.org/gsea/msigdb}
}
\usage{
evbiog
}
\description{
Extracellular vesicle biogenesis genes of \code{"Human"}, \code{"Mouse"}, and \code{"Rat"}.
}
\keyword{datasets}
