% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{find_cci}
\alias{find_cci}
\title{Infer cell-cell communications mediated by EV-derived miRNAs}
\usage{
find_cci(
  object,
  mir2tar,
  min_cell_num = 10,
  pvalue = 0.05,
  resolution = "mature",
  min_percent = 0.05,
  if_doParallel = TRUE,
  use_n_cores = 4
)
}
\arguments{
\item{object}{miRTalk object after \code{\link{create_miRTalk}}}

\item{mir2tar}{A data.frame of the system data containing relationship of miRNA and its target genes for \code{'Human'}, \code{'Mouse'} or \code{'Rat'}. see \code{\link{demo_mir2tar}}}

\item{min_cell_num}{Min cell number for each cell type and expressed miRNA. Default is \code{10}}

\item{pvalue}{Cutoff of p value. Default is \code{0.05}}

\item{resolution}{Correct to precursor or mature miRNAs. Use 'precursor' or 'mature'. Default is \code{'mature'}}

\item{min_percent}{Min percent of expressed cells for target genes of miRNA. Default is \code{0.05}}

\item{if_doParallel}{Use doParallel. Default is TRUE.}

\item{use_n_cores}{Number of CPU cores to use. Default is 4}
}
\value{
miRTalk object containing the inferred cell-cell communications mediated by EV-derived miRNAs
}
\description{
Infer cell-cell communications mediated by exosomal miRNAs from senders to receivers
}
