% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{find_hvtg}
\alias{find_hvtg}
\title{Find highly variable target genes}
\usage{
find_hvtg(
  object,
  pvalue = 0.05,
  log2fc = 0.5,
  min_cell_num = 10,
  nfeatures = 3000
)
}
\arguments{
\item{object}{miRTalk object after \code{\link{find_miRNA}}}

\item{pvalue}{Cutoff of p value. Default is \code{0.05}}

\item{log2fc}{log2 fold change for identifying the highly expressed genes in each cell type. Default is \code{0.5}}

\item{min_cell_num}{Min cell number for each cell type. Default is \code{10}}

\item{nfeatures}{Number of features to select as top variable features. Default is \code{3000}}
}
\value{
miRTalk object containing highly variable target genes without the cell-type-specific potential marker genes
}
\description{
Find highly variable target genes by excluding the cell-type-specific potential marker genes
}
