% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{find_miRNA}
\alias{find_miRNA}
\title{Find expressed miRNAs}
\usage{
find_miRNA(object, mir_info)
}
\arguments{
\item{object}{miRTalk object after \code{\link{create_miRTalk}}}

\item{mir_info}{A data.frame of the system data containing information of EV-derived miRNA of \code{'Human'}, \code{'Mouse'} or \code{'Rat'}. see \code{\link{demo_mir_info}}}
}
\value{
miRTalk object containing the expressed miRNAs
}
\description{
Find expressed miRNAs among all cells
}
