% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{find_miRTalk}
\alias{find_miRTalk}
\title{Infer cell-cell communications mediated by EV-derived miRNAs}
\usage{
find_miRTalk(
  object,
  min_cell_num = 10,
  min_percent = 0.05,
  pvalue = 0.05,
  per_num = 1000,
  if_filter_miRNA = FALSE,
  if_consider_condition = TRUE,
  if_doParallel = TRUE,
  use_n_cores = 4
)
}
\arguments{
\item{object}{miRTalk object after \code{\link{create_miRTalk}}}

\item{min_cell_num}{Min cell number for each cell type and expressed miRNA. Default is \code{10}}

\item{min_percent}{Min percent of expressed cells for target genes of miRNA. Default is \code{0.05}}

\item{pvalue}{Cutoff of p value. Default is \code{0.05}}

\item{if_filter_miRNA}{Whether to filter the significantly highly expressed miRNAs. Default is FALSE}

\item{if_consider_condition}{Whether to infer for each condition. Default is TRUE}

\item{if_doParallel}{Use doParallel. Default is TRUE}

\item{use_n_cores}{Number of CPU cores to use. Default is 4}
}
\value{
miRTalk object containing the inferred cell-cell communications mediated by EV-derived miRNAs
}
\description{
Infer cell-cell communications mediated by exosomal miRNAs from senders to receivers
}
