% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{find_miRTalk_bulk}
\alias{find_miRTalk_bulk}
\title{Infer EV-derived miR-target interactions for bulk data}
\usage{
find_miRTalk_bulk(
  rna_data,
  mirna_data,
  type,
  resolution = "mature",
  species,
  mir_info,
  mir2tar,
  if_normalize = TRUE,
  if_use_evbiog_risc = TRUE,
  evbiog = NULL,
  risc = NULL,
  score_scale_method = "1",
  target_scale_method = "1",
  database = "miRTarBase",
  regulation = "negative",
  if_use_human_data = FALSE,
  if_combine = TRUE,
  gene2gene = NULL
)
}
\arguments{
\item{rna_data}{RNA-seq data with rows containg genes and columns containing samples.}

\item{mirna_data}{miRNA-seq data with rows containg miRNA and columns containing samples.}

\item{type}{Which types of mirna_data, miRNA genes, precursor or mature miRNAs. Use 'gene', 'precursor' or 'mature'.}

\item{resolution}{Correct to precursor or mature miRNAs. Use 'precursor' or 'mature'. Default is \code{'mature'}}

\item{species}{A character meaning species of the single-cell transcriptomics data.\code{'Human'}, \code{'Mouse'} or \code{'Rat'}}

\item{mir_info}{A data.frame of the system data containing information of EV-derived miRNA of \code{'Human'}, \code{'Mouse'} or \code{'Rat'}. see \code{\link{demo_mir_info}}}

\item{mir2tar}{A data.frame of the system data containing relationship of miRNA and its target genes for \code{'Human'}, \code{'Mouse'} or \code{'Rat'}. see \code{\link{demo_mir2tar}}}

\item{if_normalize}{Normalize sc_data with Seurat LogNormalize. Set it \code{FLASE} when rna_data and mirna_data have been normalized.}

\item{if_use_evbiog_risc}{if considering module score of extracellular vesicle biogenesis genes and RNA-induced silencing complex related genes. Default is TRUE. Consider set it FALSE for the comparion of scores between various conditions.}

\item{evbiog}{A data.frame of the system data containing extracellular vesicle biogenesis genes of \code{"Human"}, \code{"Mouse"}, and \code{"Rat"}.}

\item{risc}{A data.frame of the system data containing RNA-induced silencing complex related genes of \code{"Human"}, \code{"Mouse"}, and \code{"Rat"}.}

\item{score_scale_method}{Methods for scale the Seurat scores of evbiog and risc signatures, "1" for min max scale used in scRNA-seq data by default, and "2" for rank scale. For small sample size, set it to "2" to reduce the zero scores.}

\item{target_scale_method}{Methods for scale the target gene expression, "1" for rank scale used in scRNA-seq data by default, "2" for the scale with values divided by the max value for each sample, "3" for min max scale. For bulk RNA-seq, consider set it to "2" (all values >= 0) or "3" to reduce the significant heterogeneity of samples, especially human samples.}

\item{database}{Which database of miRNA-target interactions to use, "miRTarBase" and/or "TarBase". Default is the "miRTarBase". It can also be "TarBase" or c("miRTarBase", "TarBase")}

\item{regulation}{Inference of negative or positive regulation. Default is "negative". Set it as "positive" and set database as "TarBase" for inferring positive regulation.}

\item{if_use_human_data}{Whether to use homologous human data in mir_info and mir2tar for mouse or rat scRNA-seq data. For human scRNA-seq data, no need to do it. For mouse or rat data, you can set it TRUE.}

\item{if_combine}{Whether to use combined homologous mir_info and mir2tar when if_use_human_data is TRUE. Default is TRUE.}

\item{gene2gene}{A data.frame of the system data containing the gene orthologs among human, mouse, and rat. If if_use_human_data is TRUE, please provide it, like "gene2gene = gene2gene"}
}
\value{
A data.frame containing score for each miRNA-target interaction acoss samples
}
\description{
Infer EV-derived miR-target interactions for paired bulk RNA-seq and miRNA-seq data
}
