% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{gene2go}
\alias{gene2go}
\title{gene2go}
\format{
An object of class \code{data.frame} with 1190681 rows and 7 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/gene}
}
\usage{
gene2go
}
\description{
A data.frame containing GO terms, i.e., process, function, and component, of \code{"Human"}, \code{"Mouse"}, and \code{"Rat"} updated on Jan, 02, 2022 for revising count matrix.
}
\keyword{datasets}
