% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{gene2path}
\alias{gene2path}
\title{gene2path}
\format{
An object of class \code{data.frame} with 2817311 rows and 5 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/gene}
}
\usage{
gene2path
}
\description{
A data.frame containing relationship of genes and related pathways, i.e., KEGG, GO_BP, REACTOME, WIKIPATHWAYS of \code{"Human"}, \code{"Mouse"}, and \code{"Rat"} updated on Jan, 02, 2022 for revising count matrix.
}
\keyword{datasets}
