% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{geneinfo}
\alias{geneinfo}
\title{geneinfo}
\format{
An object of class \code{data.frame} with 288559 rows and 3 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/gene}
}
\usage{
geneinfo
}
\description{
A data.frame containing gene symbols of \code{"Human"}, \code{"Mouse"}, and \code{"Rat"} updated on Jan, 02, 2022 for revising count matrix.
}
\keyword{datasets}
