% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_gene2go}
\alias{get_gene2go}
\title{Get GO terms}
\usage{
get_gene2go(target_genes, gene2go, species)
}
\arguments{
\item{target_genes}{Character of one or more target genes}

\item{gene2go}{A data.frame of the system data containing GO terms for \code{'Human'}, \code{'Mouse'} or \code{'Rat'}. see \code{\link{demo_gene2go}}}

\item{species}{A character meaning species of the target genes.\code{'Human'}, \code{'Mouse'} or \code{'Rat'}}
}
\value{
GO terms for one or more target genes
}
\description{
Get GO terms for target genes
}
