% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_miRNA}
\alias{get_miRNA}
\title{Get miRNA activity}
\usage{
get_miRNA(object, celltype_sender, resolution = "mature")
}
\arguments{
\item{object}{miRTalk object after \code{\link{find_miRNA}}}

\item{celltype_sender}{Name of celltype_sender. One cell type}

\item{resolution}{Correct to precursor or mature miRNAs. Use 'precursor' or 'mature'. Default is \code{'mature'}}
}
\value{
A data.frame containing all miRNA activity.
}
\description{
Get all miRNAs' activity for a given sender cell type
}
