% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_miRTalk_cci}
\alias{get_miRTalk_cci}
\title{Get miRNA activity}
\usage{
get_miRTalk_cci(object, simple = TRUE)
}
\arguments{
\item{object}{miRTalk object after \code{\link{find_miRNA}}}

\item{simple}{Whether to show the simple results. Default is \code{TRUE}}
}
\value{
A data.frame containing all miRNA activity.
}
\description{
Get all miRNAs' activity for a given sender cell type
}
