% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_miRTalk_circulating_score}
\alias{get_miRTalk_circulating_score}
\title{Get circulating score of inferred miRNAs}
\usage{
get_miRTalk_circulating_score(object)
}
\arguments{
\item{object}{miRTalk object after \code{\link{find_miRTalk}}}
}
\value{
A data.frame containing all potential circulating miRNAs.
}
\description{
Get circulating score of inferred miRNAs.
}
