% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_miRTalk_pathway}
\alias{get_miRTalk_pathway}
\title{Get overlapped pathways}
\usage{
get_miRTalk_pathway(
  object,
  gene2path,
  mir2path,
  miRNA = NULL,
  targetgenes = NULL
)
}
\arguments{
\item{object}{miRTalk object after \code{\link{find_miRTalk}}}

\item{gene2path}{A data.frame of the system data containing gene-related pathways from KEGG, Reactome, GO_BP, Wikipathways for \code{'Human'}, \code{'Mouse'} or \code{'Rat'}.}

\item{mir2path}{A data.frame of the system data containing miRNA-related pathways from KEGG, Reactome, GO_BP, Wikipathways for \code{'Human'}, \code{'Mouse'} or \code{'Rat'}.}

\item{miRNA}{which miRNAs to analyze. Default is all inferred miRNAs in senders.}

\item{targetgenes}{which targetgenes to analyze. Default is all inferred target genes in receivers.}
}
\value{
A list of pathways for miRNAs and target genes.
}
\description{
Get overlapped pathways between miRNA and target gene related pathways
}
