% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{mir2path}
\alias{mir2path}
\title{mir2path}
\format{
An object of class \code{data.frame} with 130339 rows and 4 columns.
}
\source{
\url{https://mirtarbase.cuhk.edu.cn/}
}
\usage{
mir2path
}
\description{
A data.frame containing relationship of miRNAs and related pathways, i.e., KEGG, GO_BP, REACTOME, WIKIPATHWAYS of \code{"Human"}, \code{"Mouse"}, and \code{"Rat"} updated on Jan, 02, 2022 for revising count matrix.
}
\keyword{datasets}
