% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{mir2tar}
\alias{mir2tar}
\title{mir2tar}
\format{
An object of class \code{data.frame} with 544329 rows and 8 columns.
}
\source{
\url{https://mirtarbase.cuhk.edu.cn/}
}
\usage{
mir2tar
}
\description{
A data.frame containing relationship of miRNA and its target genes for \code{"Human"}, \code{"Mouse"}, and \code{"Rat"}.
}
\keyword{datasets}
