% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{mir_info}
\alias{mir_info}
\title{mir_info}
\format{
An object of class \code{data.frame} with 3934 rows and 10 columns.
}
\source{
\url{http://bioinfo.life.hust.edu.cn/EVmiRNA#!/}

\url{https://mirtarbase.cuhk.edu.cn/}
}
\usage{
mir_info
}
\description{
A data.frame containing infomation of exosomal miRNA of \code{"Human"}, \code{"Mouse"}, and \code{"Rat"}.
}
\keyword{datasets}
