% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{miRTalk}
\alias{miRTalk}
\alias{miRTalk-class}
\title{Definition of 'miRTalk' class}
\description{
An S4 class containing the data, meta, and results of inferred cell-cell communications mediated by EV-derived miRNAs.
}
\section{Slots}{

\describe{
\item{\code{data}}{A list containing the data and variable genes.}

\item{\code{meta}}{A data.frame containing the meta data.}

\item{\code{species}}{A character containing the species.}

\item{\code{miR}}{A data.frame containing expressed miRNA genes.}

\item{\code{miR2tar}}{A data.frame containing expressed miRNAs and their target genes.}

\item{\code{type}}{A character containing the type of miRNA.}

\item{\code{per_test_list}}{Generated background distribution for permutation test.}

\item{\code{cci}}{A data.frame containing the significantly enriched EV-derived miRNAs and their target genes.}
}}

