% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{pathways}
\alias{pathways}
\title{pathways}
\format{
An object of class \code{data.frame} with 222355 rows and 5 columns.
}
\source{
\url{https://www.genome.jp/kegg/pathway.html}
}
\usage{
pathways
}
\description{
KEGG pathways of \code{"Human"}, \code{"Mouse"}, and \code{"Rat"}.
}
\keyword{datasets}
