% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cci_heatmap}
\alias{plot_cci_heatmap}
\title{Heatmap plot of cell-cell communications}
\usage{
plot_cci_heatmap(
  object,
  celltype = NULL,
  miRNA = NULL,
  show_type = "number",
  text_size = 10
)
}
\arguments{
\item{object}{miRTalk object after \code{\link{find_cci}}}

\item{celltype}{which cell types to plot by order. Default is to plot all cell types}

\item{miRNA}{which miRNAs to use. Default is to plot all inferred miRNAs}

\item{show_type}{which type of miRNAs to show, \code{"number"} and \code{"activity"} for sum of inferred miRNAs number and activity, respectively, or \code{"bayes"} for max Bayes probability. Default is \code{"number"}}

\item{text_size}{Size of text labels. Default is \code{10}}
}
\value{
Heatmap plot of cell-cell communications mediated by EV-derived miRNA
}
\description{
Heatmap plot of cell-cell communications from senders to receivers with the sum of inferred miRNAs number or activity, or the max Bayes probability
}
