% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_miR2tar_chord}
\alias{plot_miR2tar_chord}
\title{Chord plot of EV-derived miRNAs and target genes}
\usage{
plot_miR2tar_chord(
  object,
  celltype_sender,
  celltype_receiver,
  celltype_color = NULL,
  miRNA = NULL,
  edge_color = NULL,
  edge_type = "ellipse",
  text_size = 0.5
)
}
\arguments{
\item{object}{miRTalk object after \code{\link{find_cci}}}

\item{celltype_sender}{Name of celltype_sender. One or more cell types}

\item{celltype_receiver}{Name of celltype_receiver. One or more cell types}

\item{celltype_color}{Colors for the celltype_sender nodes and celltype_receiver nodes, or use \code{"NO"} to make it simple}

\item{miRNA}{which miRNAs to use. Default is to plot all inferred miRNAs}

\item{edge_color}{Colors for the edges from the sender cell type}

\item{edge_type}{Types for the edges from the sender cell type. Default is \code{"ellipse"}. \code{"big.arrow"} for big arrow}

\item{text_size}{Size of text labels. Default is \code{0.5}}
}
\value{
Chord plot of EV-derived miRNAs and target genes
}
\description{
Chord plot of EV-derived miRNAs and target genes from senders to receivers.
}
