% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_miR2tar_heatmap}
\alias{plot_miR2tar_heatmap}
\title{Heatmap plot of EV-derived miRNAs and target genes}
\usage{
plot_miR2tar_heatmap(
  object,
  celltype_sender,
  celltype_receiver,
  miRNA = NULL,
  text_size = 5,
  if_horizontal = TRUE
)
}
\arguments{
\item{object}{miRTalk object after \code{\link{find_cci}}}

\item{celltype_sender}{Name of celltype_sender}

\item{celltype_receiver}{Name of celltype_receiver}

\item{miRNA}{which miRNAs to use. Default is to plot all inferred miRNAs}

\item{text_size}{Size of text labels. Default is \code{3}}

\item{if_horizontal}{Whether to plot with the horizontal direction. Default is \code{TRUE}}
}
\value{
Heatmap plot of EV-derived miRNAs and target genes
}
\description{
Heatmap plot of EV-derived miRNAs and target genes from senders to receivers.
}
