% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_miR_bubble}
\alias{plot_miR_bubble}
\title{Bubble plot of inferred miRNA}
\usage{
plot_miR_bubble(
  object,
  celltype = NULL,
  miRNA = NULL,
  if_show_autocrine = FALSE,
  if_horizontal = TRUE,
  viridis_option = "D"
)
}
\arguments{
\item{object}{miRTalk object after \code{\link{find_miRTalk}}}

\item{celltype}{which cell types to plot. Default is to plot all cell types}

\item{miRNA}{which miRNAs to plot. Default is to plot all inferred miRNAs}

\item{if_show_autocrine}{Whether to show autocrine. Default is \code{FALSE}}

\item{if_horizontal}{Whether to plot with the horizontal direction. Default is \code{TRUE}}

\item{viridis_option}{option in \code{\link[viridis]{scale_color_viridis}}, can be "A", "B", "C", "D", "E", "F", "G", "H". Default is "D".}
}
\value{
ggplot2 object for Bubble plot of inferred miRNA
}
\description{
Bubble plot of inferred miRNA from senders top receivers. Rows for cell pairs, and columns for miRNAs by default.
}
