% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_miR_heatmap}
\alias{plot_miR_heatmap}
\title{Heatmap plot of inferred miRNA}
\usage{
plot_miR_heatmap(
  object,
  celltype = NULL,
  miRNA = NULL,
  text_size = 10,
  if_horizontal = TRUE
)
}
\arguments{
\item{object}{miRTalk object after \code{\link{find_cci}}}

\item{celltype}{which cell types to plot. Default is to plot all cell types}

\item{miRNA}{which miRNAs to plot. Default is to plot all inferred miRNAs}

\item{text_size}{Size of text labels. Default is \code{10}}

\item{if_horizontal}{Whether to plot with the horizontal direction. Default is \code{TRUE}}
}
\value{
Heatmap plot of inferred miRNA
}
\description{
heatmap plot of inferred miRNA for each sender. Rows for cell types, and columns for miRNAs by default
}
