% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_miRTalk_chord}
\alias{plot_miRTalk_chord}
\title{Chord plot of cell-cell communications}
\usage{
plot_miRTalk_chord(
  object,
  celltype = NULL,
  celltype_color = NULL,
  miRNA = NULL,
  edge_color = NULL,
  edge_type = "big.arrow",
  show_type = "number",
  if_show_autocrine = FALSE,
  text_size = 1.5,
  y_scale = 0.1,
  ...
)
}
\arguments{
\item{object}{miRTalk object after \code{\link{find_miRTalk}}}

\item{celltype}{which cell types to plot by order. Default is to plot all cell types}

\item{celltype_color}{Colors for the cell types, whose length must be equal to \code{celltype}}

\item{miRNA}{which miRNAs to use. Default is to plot all inferred miRNAs}

\item{edge_color}{Colors for the edges from the sender cell type, whose length must be equal to \code{celltype}}

\item{edge_type}{Types for the edges from the sender cell type. Default is \code{"big.arrow"}. \code{"ellipse"} for ellipse, "triangle" for triangle, "curved" for curved. Details see \code{\link[circlize]{chordDiagram}}}

\item{show_type}{which type of miRNAs to show, \code{"number"}, \code{"activity"}, or \code{"score"} for sum of inferred miRNAs number and activity, respectively, or \code{"prob"} for max probability. Default is \code{"number"}}

\item{if_show_autocrine}{Whether to show autocrine. Default is \code{FALSE}}

\item{text_size}{Size of text labels. Default is \code{1.5}}

\item{y_scale}{y_scale to adjust the text. Default is \code{0.1}}

\item{...}{parameters pass to \code{\link[circlize]{chordDiagram}}, e.g., link.arr.width, link.arr.length, link.arr.col}
}
\value{
Chord plot of cell-cell communications mediated by EV-derived miRNA
}
\description{
Chord plot of cell-cell communications from senders to receivers with the sum of inferred miRNAs number, activity, score, or the max probability
}
