% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_miRTalk_circle}
\alias{plot_miRTalk_circle}
\title{Circle plot of cell-cell communications}
\usage{
plot_miRTalk_circle(
  object,
  celltype = NULL,
  miRNA = NULL,
  celltype_color = NULL,
  edge_color = NULL,
  edge_type = "fan",
  show_type = "number",
  if_show_autocrine = FALSE,
  edge_alpha = 0.5,
  node_size = 10,
  text_size = 5
)
}
\arguments{
\item{object}{miRTalk object after \code{\link{find_miRTalk}}}

\item{celltype}{which cell types to plot. Default is to plot all cell types}

\item{miRNA}{which miRNAs to use. Default is to plot all inferred miRNAs}

\item{celltype_color}{Colors for the cell types, whose length must be equal to \code{celltype}}

\item{edge_color}{Colors for the edges from the sender cell type, whose length must be equal to \code{celltype}}

\item{edge_type}{Types for the edges. \code{"fan"} by default, \code{"link"}, \code{"hive"}}

\item{show_type}{which type of miRNAs to show, \code{"number"}, \code{"activity"}, or \code{"score"} for sum of inferred miRNAs number and activity, respectively, or \code{"prob"} for max probability. Default is \code{"number"}}

\item{if_show_autocrine}{Whether to show autocrine. Default is \code{FALSE}}

\item{edge_alpha}{Transparency of edge. Default is \code{0.5}}

\item{node_size}{Size of node. Default is \code{10}}

\item{text_size}{Size of text. Default is \code{5}}
}
\value{
ggplot2 object for Circle plot of cell-cell communications mediated by EV-derived miRNA
}
\description{
Circle plot of cell-cell communications from senders to receivers with the sum of inferred miRNAs number, activity, score, or the max probability
}
