% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_miRTalk_circle_simple}
\alias{plot_miRTalk_circle_simple}
\title{Circle plot of cell-cell communications by retaining all cell type nodes}
\usage{
plot_miRTalk_circle_simple(
  object,
  condition = NULL,
  celltype,
  celltype_dir = "sender",
  miRNA = NULL,
  celltype_color = NULL,
  edge_color = NULL,
  edge_type = "fan",
  show_type = "number",
  if_show_autocrine = FALSE,
  edge_alpha = 0.5,
  node_size = 10,
  text_size = 5
)
}
\arguments{
\item{object}{miRTalk object after \code{\link{find_miRTalk}}}

\item{condition}{which conditions to plot. Default is plot all conditions.}

\item{celltype}{which cell types to plot. one or more cell types}

\item{celltype_dir}{which direction to plot, \code{"sender"} or \code{"receiver"}. Default is as \code{"sender"}.}

\item{miRNA}{which miRNAs to use. Default is to plot all inferred miRNAs in senders.}

\item{celltype_color}{Colors for the cell types, whose length must be equal to \code{celltype}}

\item{edge_color}{Colors for the edges from the sender cell type, whose length must be equal to \code{celltype}}

\item{edge_type}{Types for the edges. \code{"fan"} by default, \code{"link"}, \code{"hive"}}

\item{show_type}{which type of miRNAs to show, \code{"number"}, \code{"EVmiR_score"}, or \code{"score"} for sum of inferred miRNAs number, EVmiR_score, and MiTI_score, respectively. Default is \code{"number"}}

\item{if_show_autocrine}{Whether to show autocrine. Default is \code{FALSE}}

\item{edge_alpha}{Transparency of edge. Default is \code{0.5}}

\item{node_size}{Size of node. Default is \code{10}}

\item{text_size}{Size of text. Default is \code{5}}
}
\value{
ggplot2 object for Circle plot of cell-cell communications mediated by EV-derived miRNA
}
\description{
Circle plot of cell-cell communications from senders to receivers with the sum of inferred miRNAs number, EVmiR_score, or score by retaining all cell type nodes
}
