% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_miRTalk_heatmap}
\alias{plot_miRTalk_heatmap}
\title{Heatmap plot of cell-cell communications}
\usage{
plot_miRTalk_heatmap(
  object,
  celltype = NULL,
  miRNA = NULL,
  show_type = "number",
  text_size = 10,
  viridis_option = "D",
  ...
)
}
\arguments{
\item{object}{miRTalk object after \code{\link{find_miRTalk}}}

\item{celltype}{which cell types to plot by order. Default is to plot all cell types}

\item{miRNA}{which miRNAs to use. Default is to plot all inferred miRNAs}

\item{show_type}{which type of miRNAs to show, \code{"number"}, \code{"activity"}, or \code{"score"} for sum of inferred miRNAs number and activity, respectively, or \code{"prob"} for max probability. Default is \code{"number"}}

\item{text_size}{Size of text labels. Default is \code{10}}

\item{viridis_option}{option in \code{\link[viridis]{scale_color_viridis}}, can be "A", "B", "C", "D", "E", "F", "G", "H". Default is "D".}

\item{...}{parameters pass to \code{\link[heatmaply]{heatmaply}}, e.g., grid_color, grid_width}
}
\value{
Heatmap plot of cell-cell communications mediated by EV-derived miRNA
}
\description{
Heatmap plot of cell-cell communications from senders to receivers with the sum of inferred miRNAs number, activity, score, or the max probability
}
