% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_target_heatmap}
\alias{plot_target_heatmap}
\title{Heatmap plot of inferred targets in receivers}
\usage{
plot_target_heatmap(
  object,
  condition = NULL,
  celltype,
  targetgenes = NULL,
  limits = c(0, 1),
  text_size = 10,
  if_horizontal = TRUE,
  viridis_option = "D",
  ...
)
}
\arguments{
\item{object}{miRTalk object after \code{\link{find_miRTalk}}}

\item{condition}{which conditions to plot. Default is plot all conditions.}

\item{celltype}{which cell types as receivers to plot, one or more cell types.}

\item{targetgenes}{which targetgenes to plot. Default is to plot all inferred target genes in receivers.}

\item{limits}{A parameter \code{\link[heatmaply]{heatmaply}}, a two dimensional numeric vector specifying the data range for the scale. Default is 0-1}

\item{text_size}{Size of text labels. Default is \code{10}}

\item{if_horizontal}{Whether to plot with the horizontal direction. Default is \code{TRUE}}

\item{viridis_option}{option in \code{\link[viridis]{scale_color_viridis}}, can be "A", "B", "C", "D", "E", "F", "G", "H". Default is "D".}

\item{...}{parameters pass to \code{\link[heatmaply]{heatmaply}}, e.g., grid_color, grid_width}
}
\value{
Heatmap plot of inferred targets
}
\description{
heatmap plot of inferred targets in receivers. Rows for cell-type-specific miRNAs, and columns for targets in receivers by default
}
