% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_gene.R
\name{rev_gene}
\alias{rev_gene}
\title{Pre-processing step: revising gene symbols}
\usage{
rev_gene(data = NULL, data_type = NULL, species = NULL, geneinfo = NULL)
}
\arguments{
\item{data}{A data.frame or matrix or dgCMatrix}

\item{data_type}{A character to define the type of \code{data}, select \code{'count'} for the data matrix, \code{'mir_info'} for the mir_info data.frame, \code{'mir2tar'} for the mir2tar data.frame, \code{'pathways'} for the pathways data.frame, \code{'GO_BP'} for the GO_BP data.frame}

\item{species}{Species of the data.\code{'Human'}, \code{'Mouse'} or \code{'Rat'}}

\item{geneinfo}{A data.frame of the system data containing gene symbols of \code{'Human'}, \code{'Mouse'} and \code{'Rat'} updated on June 19, 2022 for revising gene symbols}
}
\value{
A new data.frame, matrix, or dgCMatrix.
}
\description{
Revise genes according to NCBI Gene symbols updated in June 19, 2022 for count matrix, user-custom mir_info data.frame, and mir2tar data.frame
}
