% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{risc}
\alias{risc}
\title{risc}
\format{
An object of class \code{data.frame} with 36 rows and 2 columns.
}
\source{
\url{https://doi.org/10.1016/j.molcel.2021.11.026}
}
\usage{
risc
}
\description{
RNA-induced silencing complex related genes of \code{"Human"}, \code{"Mouse"}, and \code{"Rat"}.
}
\keyword{datasets}
