% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{ritac}
\alias{ritac}
\title{ritac}
\format{
An object of class \code{data.frame} with 45 rows and 2 columns.
}
\source{
\url{https://doi.org/10.1038/cr.2016.22}
}
\usage{
ritac
}
\description{
RNA-induced transcriptional activation complex related genes of \code{"Human"}, \code{"Mouse"}, and \code{"Rat"}.
}
\keyword{datasets}
