% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_miR_bubble}
\alias{plot_miR_bubble}
\title{Bubble plot of inferred miRNA}
\usage{
plot_miR_bubble(
  object,
  celltype = NULL,
  miRNA = NULL,
  if_show_autocrine = FALSE,
  if_horizontal = TRUE
)
}
\arguments{
\item{object}{miRTalk object after \code{\link{find_cci}}}

\item{celltype}{which cell types to plot. Default is to plot all cell types}

\item{miRNA}{which miRNAs to plot. Default is to plot all inferred miRNAs}

\item{if_show_autocrine}{Whether to show autocrine. Default is \code{FALSE}}

\item{if_horizontal}{Whether to plot with the horizontal direction. Default is \code{TRUE}}
}
\value{
Bubble plot of inferred miRNA
}
\description{
Bubble plot of inferred miRNA from senders top receivers. Rows for cell pairs, and columns for miRNAs by default.
}
